pyproject.toml
[project]
name = "yt-watchdog-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "sanero",email = "e.eliseev@automacon.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram (>=3.20.0.post0,<4.0.0)",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "apscheduler>=3.10.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
